requirements.txt
# Direct dependencies only — transitive deps resolved by pip
# macOS-only deps (pyobjc-*, pywebview) are in requirements-macos.txt

# AI / Agent framework
google-adk>=1.24.0
google-genai>=1.0.0
openai>=1.0.0
litellm>=1.50.0
mem0ai>=1.0.0
chromadb>=1.0.0

# Web framework & server
fastapi>=0.100.0
uvicorn>=0.23.0
sse-starlette>=1.0.0

# Database
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0
greenlet>=3.0.0

# HTTP
httpx>=0.24.0

# Email gateway (EWS)
exchangelib>=5.0.0

# Document parsing
PyMuPDF>=1.23.0
pymupdf4llm>=0.2.0

# File watching
watchdog>=3.0.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
opencc-python-reimplemented>=0.1.7
exchangelib>=5.0.0

# Testing
pytest>=7.0.0
